+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- URI handling no longer waits for confirmation to use latest build if the provided build number is too old
+
 ## 0.7.1 - 2025-10-13
 
 ### Fixed